+++ autd3-firmware-emulator/tests/op/clear.rs
@@ -3,7 +3,6 @@ use std::num::{NonZeroU16, NonZeroU8};
 use autd3_derive::Modulation;
 use autd3_driver::{
     datagram::*,
-    defined::ULTRASOUND_PERIOD,
     derive::*,
     firmware::{
         cpu::TxDatagram,
@@ -44,7 +43,7 @@ fn send_clear() -> anyhow::Result<()> {
     let mut tx = TxDatagram::new(geometry.num_devices());
 
     {
-        let d = Silencer::from_completion_time(ULTRASOUND_PERIOD, ULTRASOUND_PERIOD);
+        let d = Silencer::disable();
         assert_eq!(Ok(()), send(&mut cpu, d, &geometry, &mut tx));
 
         let d = unsafe {
@@ -83,11 +82,11 @@ fn send_clear() -> anyhow::Result<()> {
     assert_eq!(1, cpu.fpga().silencer_update_rate_intensity());
     assert_eq!(1, cpu.fpga().silencer_update_rate_phase());
     assert_eq!(
-        SILENCER_STEPS_INTENSITY_DEFAULT as u8,
+        SILENCER_STEPS_INTENSITY_DEFAULT,
         cpu.fpga().silencer_completion_steps_intensity()
     );
     assert_eq!(
-        SILENCER_STEPS_PHASE_DEFAULT as u8,
+        SILENCER_STEPS_PHASE_DEFAULT,
         cpu.fpga().silencer_completion_steps_phase()
     );
     assert!(cpu.fpga().silencer_fixed_completion_steps_mode());