+++ autd3-firmware-emulator/tests/op/modulation.rs
@@ -1,4 +1,7 @@
-use std::{num::NonZeroU16, time::Duration};
+use std::{
+    num::{NonZeroU16, NonZeroU8},
+    time::Duration,
+};
 
 use autd3_driver::{
     datagram::{IntoDatagramWithSegmentTransition, Silencer, SwapSegment},
@@ -173,10 +176,7 @@ fn mod_freq_div_too_small() -> anyhow::Result<()> {
         .with_segment(Segment::S0, Some(TransitionMode::Immediate));
         assert_eq!(Ok(()), send(&mut cpu, d, &geometry, &mut tx));
 
-        let d = Silencer::from_completion_time(
-            Silencer::DEFAULT_COMPLETION_TIME_INTENSITY,
-            Silencer::DEFAULT_COMPLETION_TIME_PHASE,
-        );
+        let d = Silencer::default();
         assert_eq!(Ok(()), send(&mut cpu, d, &geometry, &mut tx));
 
         let d = TestModulation {
@@ -189,9 +189,9 @@ fn mod_freq_div_too_small() -> anyhow::Result<()> {
         .with_segment(Segment::S1, None);
         assert_eq!(Ok(()), send(&mut cpu, d, &geometry, &mut tx));
 
-        let d = Silencer::from_completion_time(
-            Silencer::DEFAULT_COMPLETION_TIME_PHASE * 2,
-            Silencer::DEFAULT_COMPLETION_TIME_PHASE,
+        let d = Silencer::from_completion_steps(
+            NonZeroU8::new(SILENCER_STEPS_PHASE_DEFAULT * 2).unwrap(),
+            NonZeroU8::new(SILENCER_STEPS_PHASE_DEFAULT).unwrap(),
         );
         assert_eq!(Ok(()), send(&mut cpu, d, &geometry, &mut tx));